Let $RUNNER_USER run update-ca-certificates(8) and apt-get(8) (#272)
We're planning to support Git servers running with a self-signed
certificate and will introduce the way to trust such a certificate by
executing update-ca-certificates(8) at runtime.

In order to let `$RUNNER_USER` run this command, I want to add sudoers(5)
for the command.

In addition to allowing update-ca-certificates(8), apt-get(8) should be
also allowed to execute. So, I added the permission for the command,
too.